docs: add filteredstream examples
