+++ example_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"time"
 
 	"github.com/sivchari/gotwtr"
 )
@@ -78,14 +79,80 @@ func ExampleClient_CountsRecentTweet() {
 		fmt.Println(t)
 	}
 }
-func ExampleClient_AddOrDeleteRules() {
+func ExampleClient_AddOrDeleteRules_add() {
+	client := gotwtr.New("key")
+	_, err := client.AddOrDeleteRules(context.Background(), &gotwtr.AddOrDeleteJSONBody{
+		Add: []*gotwtr.AddRule{
+			{
+				Value: "puppy has:media",
+				Tag:   "puppies with media",
+			},
+			{
+				Value: "meme has:images",
+			},
+		},
+	})
+	if err != nil {
+		log.Fatal(err)
+	}
+}
 
+func ExampleClient_AddOrDeleteRules_delete() {
+	client := gotwtr.New("key")
+	// retrieve Stream rules
+	ts, err := client.RetrieveStreamRules(context.Background())
+	if err != nil {
+		panic(err)
+	}
+	var ids []string
+	for _, t := range ts.Rules {
+		fmt.Println(t)
+		ids = append(ids, t.ID)
+	}
+
+	// delete Stream rules
+	_, err = client.AddOrDeleteRules(context.Background(), &gotwtr.AddOrDeleteJSONBody{
+		Delete: &gotwtr.DeleteRule{
+			IDs: ids,
+		},
+	})
+	if err != nil {
+		panic(err)
+	}
 }
 func ExampleClient_RetrieveStreamRules() {
-
+	client := gotwtr.New("key")
+	ts, err := client.RetrieveStreamRules(context.Background())
+	if err != nil {
+		log.Fatal(err)
+	}
+	for _, t := range ts.Rules {
+		fmt.Println(t)
+	}
 }
 func ExampleClient_ConnectToStream() {
-
+	client := gotwtr.New("key")
+	ch := make(chan gotwtr.ConnectToStreamResponse, 5)
+	errCh := make(chan error)
+	stream := client.ConnectToStream(context.Background(), ch, errCh)
+	fmt.Println("streaming...")
+	ctx, cancel := context.WithCancel(context.Background())
+	go func(ctx context.Context) {
+		for {
+			select {
+			case data := <-ch:
+				fmt.Println(data.Tweet)
+			case err := <-errCh:
+				fmt.Println(err)
+			case <-ctx.Done():
+				break
+			}
+		}
+	}(ctx)
+	time.Sleep(time.Second * 10)
+	cancel()
+	stream.Stop()
+	fmt.Println("done")
 }
 func ExampleClient_VolumeStreams() {
